requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
attrs==23.1.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloud-sql-python-connector==1.4.3
colorama==0.4.6
cryptography==41.0.6
Flask==3.0.0
frozenlist==1.4.0
google-auth==2.23.4
greenlet==3.0.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyMySQL==1.1.0
python-dotenv==1.0.0
requests==2.31.0
rsa==4.9
SQLAlchemy==2.0.23
typing_extensions==4.8.0
urllib3==2.1.0
waitress==2.1.2
Werkzeug==3.0.1
yarl==1.9.3
